IGNITE-25383 Fix snapshot creation if DataRegionConfiguration#storagePath set
